refactor(agents): extract runtime-state seam from agents.ts
Move AgentRuntimeState, AgentResolution, getAgentRuntimeStateSync,
getAgentRuntimeState, saveAgentRuntimeState, patchRuntimeJson, and helpers to
src/lib/agents/runtime-state.ts; re-export from agents.ts.

Behavior-preserving move only; no logic changes.
